requirements.txt
cffi==1.11.2
cryptography==1.7.2
Django==1.11.7
django-jwt-auth==0.0.2
graphene==2.0
graphene-django==2.0.0
graphql-core==2.0
graphql-relay==0.4.5
gunicorn==19.7.1
idna==2.6
iso8601==0.1.12
promise==2.1
psycopg2==2.7.3.2
pyasn1==0.3.7
pycparser==2.18
PyJWT==1.5.3
pytz==2017.3
Rx==1.6.0
singledispatch==3.4.0.3
six==1.11.0
typing==3.6.2
